Created new scenes and folders and imported new assets
I created a new scene for the Ruby's Adventure project and saved it as a new scene file. I then imported the character Ruby as an asset in my art folder. After I made her into a sprite/game object and centered the camera and positioned Ruby in the scene window using the inspector window. Finally I made a new script folder for controlling Ruby.